+++ internal/test-utils/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @nifrajs/test-utils
 
+## 1.1.0
+
 ## 1.0.0
 
 ## 1.0.0-beta.4